rename group component to collection_details
