+++ cohort_manager/static/cohort_manager/js/components/collection_details.vue
